Remove email infra, clean DB, add icon buttons and theme improvements across all pages

+++ backend/src/main/java/com/sinte/backend/api/v1/auth/AuthController.java
@@ -74,7 +74,7 @@ public class AuthController {
     ) {
         java.util.UUID userId = SecurityUtils.currentUserId();
         List<UserPositionService.PositionAssignment> serviceAssignments = assignments.stream()
-                .map(a -> new UserPositionService.PositionAssignment(a.positionCode(), a.priority()))
+                .map(a -> new UserPositionService.PositionAssignment(a.positionCode(), (short) a.priority()))
                 .toList();
         return ResponseEntity.ok(userPositionService.setUserPositions(userId, serviceAssignments));
     }
